requirements.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
coverage==5.5
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.6
django-allauth==0.45.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
flake8==3.9.2
flake8-isort==4.0.0
idna==3.2
isort==5.9.3
mccabe==0.6.1
oauthlib==3.1.1
psycopg2==2.9.1
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
sqlparse==0.4.1
testfixtures==6.18.0
toml==0.10.2
urllib3==1.26.6
